Update to use int64 instead of int

+++ backend/internal/model/common/investment_tier.go
@@ -4,6 +4,6 @@ type InvestmentTier struct {
 	Name       string  `json:"name"`
 	MinAmount  uint64  `json:"min_amount"`
 	Multiplier float64 `json:"multiplier"`
-	PitchID    int     `json:"pitch_id"`
-	ID         *int    `json:"id,omitempty"`
+	PitchID    int64   `json:"pitch_id"`
+	ID         *int64  `json:"id,omitempty"`
 }
