+++ backend/internal/model/common/profile.go
@@ -4,6 +4,6 @@ type Profile struct {
 	ID               string `json:"id"`
 	Role             string `json:"role"`
 	DisplayName      string `json:"display_name"`
-	DashboardBalance *int   `json:"dashboard_balance,omitempty"`
+	DashboardBalance *int64 `json:"dashboard_balance,omitempty"`
 	Email            string `json:"email"`
 }
