+++ backend/internal/model/frontend/pitch_media.go
@@ -1,9 +1,9 @@
 package frontend
 
 type PitchMedia struct {
-	ID                 *int   `json:"id,omitempty"`
-	PitchID            *int   `json:"pitch_id,omitempty"`
+	ID                 *int64 `json:"id,omitempty"`
+	PitchID            *int64 `json:"pitch_id,omitempty"`
 	URL                string `json:"url"`
 	MediaType          string `json:"media_type"`
-	OrderInDescription int    `json:"order_in_description"`
+	OrderInDescription int64  `json:"order_in_description"`
 }
